requirements.txt
aiofiles==24.1.0
altgraph==0.17.4
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.4
et_xmlfile==2.0.0
Flask==3.1.1
ftputil==5.1.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pefile==2023.2.7
priority==2.0.0
pycparser==2.22
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
Quart==0.20.0
requests==2.32.4
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
